Implemented lending reserve data query
