+++ src/utils/promisedType.ts
@@ -1,0 +1,1 @@
+export type PromisedType<T> = T extends PromiseLike<infer P> ? P : never;
