+++ src/views/Markets/MarketTable.tsx
@@ -14,14 +14,13 @@ import { useAppWeb3 } from "../../hooks/appWeb3";
 import { imagesBySymbol } from "../../utils/constants";
 import { internalAddresses } from "../../utils/contracts/contractAddresses/internalAddresses";
 import { Web3Provider } from "@ethersproject/providers";
-
-export type PromisedType<T> = T extends PromiseLike<infer P> ? P : never;
+import { LendingReserveData, reserveDataFromWeb3Result } from "../../queries/lendingReserveData";
 
 export interface UseMarketDataDto {
   address: string | null | undefined;
   library: Web3Provider | undefined;
   assets: ReadonlyArray<string>;
-  reserves: Readonly<Record<string, ReserveData>>;
+  reserves: Readonly<Record<string, LendingReserveData>>;
   assetInfo: Readonly<Record<string, AssetData>>;
   marketDataQueryKey: readonly [
     "marketData",
@@ -36,79 +35,6 @@ export interface AssetData {
   symbol: string;
   address: string;
 }
-
-export interface ReserveData {
-  //stores the reserve configuration
-  configuration: [BigNumber & ReserveConfigurationMap] & { data: BigNumber };
-  //the liquidity index. Expressed in ray
-  liquidityIndex: FixedNumber;
-  //variable borrow index. Expressed in ray
-  variableBorrowIndex: FixedNumber;
-  //the current supply rate. Expressed in ray
-  currentLiquidityRate: FixedNumber;
-  //the current variable borrow rate. Expressed in ray
-  currentVariableBorrowRate: FixedNumber;
-  //the current stable borrow rate. Expressed in ray
-  currentStableBorrowRate: FixedNumber;
-  lastUpdateTimestamp: number;
-  //tokens addresses
-  aTokenAddress: string;
-  stableDebtTokenAddress: string;
-  variableDebtTokenAddress: string;
-  //address of the interest rate strategy
-  interestRateStrategyAddress: string;
-  //the id of the reserve. Represents the position in the list of the active reserves
-  id: number;
-}
-
-type Web3ReserveDataResult = PromisedType<
-  ReturnType<typeof AgaveLendingABI.prototype.getReserveData>
->;
-
-export function reserveDataFromWeb3Result({
-  configuration,
-  liquidityIndex,
-  variableBorrowIndex,
-  currentLiquidityRate,
-  currentVariableBorrowRate,
-  currentStableBorrowRate,
-  lastUpdateTimestamp,
-  aTokenAddress,
-  stableDebtTokenAddress,
-  variableDebtTokenAddress,
-  interestRateStrategyAddress,
-  id,
-}: Web3ReserveDataResult): ReserveData {
-  return {
-    configuration,
-    liquidityIndex: FixedNumber.fromValue(liquidityIndex, 27),
-    variableBorrowIndex: FixedNumber.fromValue(variableBorrowIndex, 27),
-    currentLiquidityRate: FixedNumber.fromValue(currentLiquidityRate, 27),
-    currentVariableBorrowRate: FixedNumber.fromValue(
-      currentVariableBorrowRate,
-      27
-    ),
-    currentStableBorrowRate: FixedNumber.fromValue(currentStableBorrowRate, 27),
-    lastUpdateTimestamp,
-    aTokenAddress,
-    stableDebtTokenAddress,
-    variableDebtTokenAddress,
-    interestRateStrategyAddress,
-    id,
-  };
-}
-
-//bit 0-15: LTV
-//bit 16-31: Liq. threshold
-//bit 32-47: Liq. bonus
-//bit 48-55: Decimals
-//bit 56: Reserve is active
-//bit 57: reserve is frozen
-//bit 58: borrowing is enabled
-//bit 59: stable rate borrowing enabled
-//bit 60-63: reserved
-//bit 64-79: reserve factor
-export interface ReserveConfigurationMap extends BigNumber {}
 
 export const useMarketData = (): UseMarketDataDto => {
   const { account: address, library, chainId } = useAppWeb3();
@@ -175,7 +101,7 @@ export const useMarketData = (): UseMarketDataDto => {
       console.log(reserveMap, assetInfo);
       return {
         assets: reserveList,
-        reserves: reserveMap as Record<string, ReserveData>,
+        reserves: reserveMap as Record<string, LendingReserveData>,
         assetInfo,
       };
     },
